backend/requirements.txt
fastapi 
pymongo[srv]
python-dotenv
pydantic
uvicorn
asyncio

pyjwt
pwdlib[argon2]
python-multipart

